+++ src/main/java/com/wildcodeschool/giftmefive/controller/signUpController.java
@@ -25,7 +25,7 @@ public class signUpController {
 
     @PostMapping("/enregistre_user")
     public String userUpdate(Model out,
-                             @RequestParam int idUser,
+                             @RequestParam Long idUser,
                              @RequestParam String username,
                              @RequestParam String password,
                              @RequestParam String passwordConfirmation,
