+++ src/main/java/com/wildcodeschool/giftmefive/repository/ListsRepository.java
@@ -23,12 +23,12 @@ public class ListsRepository {
             ResultSet resultSet = statement.executeQuery();
             List<ListGift> lists = new ArrayList<ListGift>();
             while (resultSet.next()) {
-                int idList = resultSet.getInt("id_list");
+                Long idList = resultSet.getLong("id_list");
                 String listName = resultSet.getString("list_name");
                 String description = resultSet.getString("description");
                 String urlImage = resultSet.getString("url_image");
                 String urlShare = resultSet.getString("url_share");
-                int idUser = resultSet.getInt("id_user");
+                long idUser = resultSet.getLong("id_user");
                 lists.add(new ListGift(idList, listName, description, urlImage, urlShare, idUser));
             }
             return lists;
@@ -37,7 +37,7 @@ public class ListsRepository {
         }
         return null;
     }
-    public ListGift findById(int idList) {
+    public ListGift findById(Long idList) {
         try {
             Connection connection = DriverManager.getConnection(
                     DB_URL, DB_USER, DB_PASSWORD
@@ -53,14 +53,14 @@ public class ListsRepository {
             String description = resultSet.getString("description");
             String urlImage = resultSet.getString("url_image");
             String urlShare = resultSet.getString("url_share");
-            int idUser = resultSet.getInt("id_user");
+            Long idUser = resultSet.getLong("id_user");
             return new ListGift(idList, listName, description, urlImage, urlShare, idUser);
         } catch (SQLException e) {
             e.printStackTrace();
         }
         return null;
     }
-        public void deleteList(int id) {
+        public void deleteList(Long id) {
             try {
                 Connection connection = DriverManager.getConnection(
                         DB_URL, DB_USER, DB_PASSWORD
@@ -68,7 +68,7 @@ public class ListsRepository {
                 PreparedStatement statement = connection.prepareStatement(
                         "DELETE FROM list WHERE id_list=?;"
             );
-            statement.setInt(1, id);
+            statement.setLong(1, id);
             if (statement.executeUpdate() != 1) {
                 throw new SQLException("failed to delete data");
             }
@@ -76,7 +76,7 @@ public class ListsRepository {
             e.printStackTrace();
         }
     }
-    public void deleteGiftFromList(int id) {
+    public void deleteGiftFromList(Long id) {
         try {
             Connection connection = DriverManager.getConnection(
                     DB_URL, DB_USER, DB_PASSWORD
@@ -84,7 +84,7 @@ public class ListsRepository {
             PreparedStatement statement = connection.prepareStatement(
                     "DELETE FROM gift WHERE id_list=?;"
             );
-            statement.setInt(1, id);
+            statement.setLong(1, id);
             if (statement.executeUpdate() != 1) {
                 throw new SQLException("failed to delete data");
             }
@@ -92,7 +92,7 @@ public class ListsRepository {
             e.printStackTrace();
         }
     }
-    public void updateList(int idList, String listName, String description, String urlImage) {
+    public void updateList(Long idList, String listName, String description, String urlImage) {
         try {
             Connection connection = DriverManager.getConnection(
                     DB_URL, DB_USER, DB_PASSWORD
@@ -103,7 +103,7 @@ public class ListsRepository {
             statement.setString(1, listName);
             statement.setString(2, description);
             statement.setString(3, urlImage);
-            statement.setInt(4, idList);
+            statement.setLong(4, idList);
             if (statement.executeUpdate() != 1) {
                 throw new SQLException("failed to update data");
             }
